Extract tests for isInt32Array to own file
